refactor: section 별 컴포넌트 분할
