+++ src/components/pages/client/calendar.tsx
@@ -1,16 +1,14 @@
-'use client';
-
-import { useState } from 'react';
-
 import { DayPicker } from 'react-day-picker';
 import { StyleWrapper } from './calendar.style';
 import 'react-day-picker/dist/style.css';
 
-export default function Calendar() {
-  const [selectedDate, setSelectedDate] = useState<Date | undefined>(
-    new Date(),
-  );
-
+export default function Calendar({
+  date,
+  onChange,
+}: {
+  date: Date | undefined;
+  onChange: React.Dispatch<React.SetStateAction<Date | undefined>>;
+}) {
   return (
     <StyleWrapper>
       <DayPicker
@@ -20,11 +18,11 @@ export default function Calendar() {
         toYear={2024}
         showOutsideDays
         required
-        selected={selectedDate}
+        selected={date}
         modifiers={{ booked: [new Date(2024, 1, 11)] }}
         modifiersStyles={{ booked: { border: '2px solid currentColor' } }}
         weekStartsOn={1}
-        onSelect={setSelectedDate}
+        onSelect={onChange}
       />
     </StyleWrapper>
   );
