+++ src/pages/client/[id]/index.tsx
@@ -1,46 +1,10 @@
-import { Divider, Flex, Heading } from '@chakra-ui/react';
+import { Flex, Heading } from '@chakra-ui/react';
 
 import GNBLayout from '@/layouts/gnb-layout';
 import DashboardItem from '@/components/common/dashboard-item';
-import Card from '@/components/pages/client/card';
 import Profile from '@/components/pages/client/profile';
-import Calendar from '@/components/pages/client/calendar';
-import { DividerWithContent } from '@/components/common/divider-with-content';
-
-const MOCK_DATA = [
-  {
-    from: '조성원',
-    to: '김도경',
-    type: '감사',
-    time: new Date(),
-    preview:
-      '편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다.',
-  },
-  {
-    from: '신승준',
-    to: '김도경',
-    type: '인사',
-    time: new Date(),
-    preview:
-      '편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다.',
-  },
-  {
-    from: '조성원',
-    to: '신승준',
-    type: '응원',
-    time: new Date(),
-    preview:
-      '편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다.',
-  },
-  {
-    from: '김도경',
-    to: '신승준',
-    type: '축하',
-    time: new Date(),
-    preview:
-      '편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다. 편지 내용이 미리보기로 나타납니다.',
-  },
-];
+import CardSection from '@/components/pages/client/card-section';
+import BirthdaySection from '@/components/pages/client/birthday-section';
 
 export default function Home() {
   return (
@@ -62,18 +26,10 @@ export default function Home() {
           w="full"
           h="full"
         >
-          <DividerWithContent
-            content="12월 20일"
-            bg="var(--chakra-colors-waggle-gray-100)"
-            fontSize={12}
-          />
-          {MOCK_DATA.map((data, index) => (
-            <Card key={index} data={data} />
-          ))}
+          <CardSection />
         </DashboardItem>
         <DashboardItem flexDir="column">
-          <Calendar />
-          <Divider />
+          <BirthdaySection />
         </DashboardItem>
       </Flex>
     </GNBLayout>
